requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-groq==0.3.5
langgraph==0.5.1
python-dotenv==1.1.1
pytest==8.4.1
pytest-asyncio==1.0.0
httpx==0.28.1
